# docs/api-specification.yaml
openapi: 3.1.0
info:
  title: AI Part Designer API
  description: |
    RESTful API for AI-powered 3D part design and generation platform.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the 
    `Authorization` header as `Bearer <token>`.
    
    ## Rate Limiting
    - Free tier: 30 requests/minute
    - Pro tier: 100 requests/minute
    - Enterprise tier: 500 requests/minute
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
    
    ## Versioning
    API version is specified in the URL path: `/api/v1/`
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@aipartdesigner.com
  license:
    name: Proprietary
    url: https://aipartdesigner.com/terms
servers:
  - url: https://api.aipartdesigner.com/api/v1
    description: Production server
  - url: https://api.staging.aipartdesigner.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List user's jobs
      operationId: listJobs
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, queued, processing, completed, failed, cancelled]
        - name: type
          in: query
          schema:
            type: string
            enum: [generate, modify, export, import, thumbnail]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Jobs list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /jobs/{jobId}:
    get:
      tags: [Jobs]
      summary: Get job status
      operationId: getJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs/{jobId}/cancel:
    post:
      tags: [Jobs]
      summary: Cancel a pending job
      operationId: cancelJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Job cannot be cancelled (already completed/processing)

  # ============================================================================
  # PROJECTS
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [generate, modify, export, import, thumbnail]
        status:
          type: string
          enum: [pending, queued, processing, completed, failed, cancelled, timeout]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        progressMessage:
          type: string
        designId:
          type: string
          format: uuid
          nullable: true
        result:
          type: object
          nullable: true
          properties:
            designId:
              type: string
              format: uuid
            versionId:
              type: string
              format: uuid
            downloadUrl:
              type: string
              format: uri
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
            message:
              type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        estimatedCompletion:
          type: string
          format: date-time
          nullable: true

    # --------------------------------------------------------------------------
    # Projects
    # --------------------------------------------------------------------------
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            requestId:
              type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
...
security:
  - bearerAuth: []